requirements.txt
pyserial
joblib
scikit-learn
numpy
pandas
matplotlib
plotly